refactor(bluetooth): decompose loading UI into reusable skeleton tile widget
